Making string formatting into a fprime implementation choice
