+++ Fw/Types/CMakeLists.txt
@@ -23,6 +23,7 @@ set(MOD_DEPS
   Fw/Cfg
 )
 register_fprime_module()
+require_fprime_implementation("Fw_StringFormat")
 ### UTs ###
 set(UT_SOURCE_FILES
   "${CMAKE_CURRENT_LIST_DIR}/test/ut/ExternalSerializeBufferTest.cpp"
@@ -35,3 +36,8 @@ register_fprime_ut()
 
 # Non-test directory
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/GTest")
+
+# Create an empty interface library and use it to register an implementation
+add_library(snprintf-format INTERFACE)
+register_fprime_implementation(Fw_StringFormat snprintf-format "${CMAKE_CURRENT_LIST_DIR}/snprintf_format.cpp")
+choose_fprime_implementation(Fw_StringFormat snprintf-format "${FPRIME_PLATFORM}") # Default choice is snprintf
