+++ Fw/Types/StringBase.cpp
@@ -14,7 +14,6 @@
 #include <Fw/Types/StringType.hpp>
 #include <Fw/Types/StringUtils.hpp>
 #include <cstdarg>
-#include <cstdio>
 #include <cstring>
 
 namespace Fw {
@@ -65,33 +64,10 @@ FormatStatus StringBase::vformat(const CHAR* formatString, va_list args) {
     CHAR* us = const_cast<CHAR*>(this->toChar());
     SizeType cap = this->getCapacity();
     FW_ASSERT(us != nullptr);
-
-    // Check format string
-    if (formatString == nullptr) {
-        return FormatStatus::INVALID_FORMAT_STRING;
-    }
-    FwSizeType total_needed_size = 0;
-#if FW_USE_PRINTF_FAMILY_FUNCTIONS_IN_STRING_FORMATTING
-    // Check that the API size type fits in fprime size type
-    static_assert(std::numeric_limits<FwSizeType>::max() >= std::numeric_limits<PlatformIntType>::max(),
-            "Range of PlatformIntType does not fit within range of FwSizeType");
-    PlatformIntType total_needed_size_api = vsnprintf(us, cap, formatString, args);
-    // Check for error return, or a type overflow
-    if (total_needed_size_api < 0) {
-        return FormatStatus::OTHER_ERROR;
-    }
-    total_needed_size = static_cast<FwSizeType>(total_needed_size_api);
-#else
-    total_needed_size = StringUtils::string_length(format_string, cap);
-    *this = formatString;
-#endif
-    // Force null terminate
-    us[cap - 1] = 0;
-    // Check for overflow
-    if (total_needed_size >= cap) {
-        return FormatStatus::OVERFLOWED;
-    }
-    return FormatStatus::SUCCESS;
+    // Needed until SizeType an FwSizeType are the same
+    static_assert(std::numeric_limits<FwSizeType>::max() >= std::numeric_limits<SizeType>::max(),
+                  "String size type must fit into FwSizeType");
+    return Fw::stringFormat(us, static_cast<FwSizeType>(cap), formatString, args);
 }
 
 bool StringBase::operator!=(const StringBase& other) const {
